Dream now supports gpsd http://gpsd.berlios.de, http://www.renderlab.net/projects/wrt54g/gpsdonwindows.html The GPS support code is always included but the thread will not be created unless Dream is compiled with the USE_GPS variable. configure with --enable-gps on linux or add #define USE_GPS into GlobalDefinitions.h on Windows. The gpsd host defaults to localhost and the port to the default gpsd port. These can be overridden in Dream.ini Dev-C++ project file update to follow. Thanks to Andy Murphy @ BBC R&D for this great code.

+++ common/GPSReceiver.cpp
@@ -171,7 +171,7 @@ void CGPSReceiver::DecodeGPSDReply(string Reply)
 	//GPSD\n\r
 	//GPSD,F=1,A=?\n\r
 
-	int GPSDPos=0;
+	size_t GPSDPos=0;
 
 	while ((GPSDPos = TotalReply.find("GPSD",0)) != string::npos)
 	{
@@ -181,8 +181,8 @@ void CGPSReceiver::DecodeGPSDReply(string Reply)
 
 		while (!finished)		// while not all of message has been consumed
 		{
-			int CurrentPos = 0;
-			for (int i=0; i < TotalReply.length(); i++)
+			size_t CurrentPos = 0;
+			for (size_t i=0; i < TotalReply.length(); i++)
 			{
 				if (TotalReply[i] == ',' || TotalReply[i] == '\r' || TotalReply[i] == '\n')
 				{
